pyproject.toml
[tool.poetry]
name = "rag-bench"
version = "0.1.0"
description = "Generalized RAG system"
authors = ["You"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
langchain = "^0.1.0"
langchain-openai = {version = "^0.0.5", extras = ["openai"]}
langchain-anthropic = "^0.1.0"
openai = "^1.12.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
tiktoken = "^0.5.2"
tenacity = "^8.2.3"
injector = "^0.21.0"
llama-index-core = "^0.10.5"
llama-index-vector-stores-postgres = "^0.1.4"
pandas = "^2.0.0"
llama-index-llms-openai = "^0.1.5"
nltk = "^3.8.1"
bs4 = "^0.0.1"
sentence-transformers = "^2.2.2"
llama-cpp-python = "^0.2.41"
pgvector = "^0.2.3"
requests = "^2.31.0"
tqdm = "^4.66.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
